refactor(server): split content-search into focused SQL modules
Decompose the 1965-line content-search.ts into a re-export facade plus modules
(entity-link, visibility, params, classification, ctes, fts, list-path,
search-path). All SQL strings and parameter bindings moved verbatim; public
signatures unchanged. Validated against the search/getContent integration tests.
